Removed dependencies on gpbackup.
Made local versions of the database and OS utility functions.

Added a retry loop to wait for the test-rig sshd service to come up. The
integration tests can fail with a "no socket" error, and we perceive a
race to get that sshd service up before a test can proceed.

Signed-off-by: Larry Hamel <lhamel@pivotal.io>